Tweaked data generator, shrunk TSV file by half
